backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# ── ML ────────────────────────────────────────────────────
catboost>=1.2.0
shap>=0.45.0
scikit-learn>=1.3.0
numpy>=1.26.0
pandas>=2.0.0

# ── Data ──────────────────────────────────────────────────
openpyxl>=3.1.0

# ── LLM / AI ──────────────────────────────────────────────
openai>=1.30.0          # used as Groq SDK client

# ── Config ────────────────────────────────────────────────
python-dotenv>=1.0.0
